Fix windows link problems with SHRLIBS
Need to link to the *.lib of each *.dll which may be in a different directory.
If no such *.lib exist, don't link but install the .dll anyway.
If only *.lib is specified, link but don't look for a corresponding *.dll,
assuming the *.dll is already installed on the target system.
Install only own *.lib and that only if this is a depencency to other modules
which is detected by having at least one header file to install.